+++ TODO.md
@@ -1,8 +1,10 @@
 ### TODO
 
 - [X] Update rosters on player signings and releases
+- [X] Swap team commands for certain roster commands
+- [X] Add possibility to hide trialing roles in team roster
 - [ ] Add coaching features
-  - [ ] Add command to create 1 on 1 coaching threads
+  - [X] Add command to create 1 on 1 coaching threads
   - [ ] Add possibility to call command by typing `LEGO® Batman™: The Videogame` in chat
   - [ ] Add command to list all the threads
   - [ ] Add command to manage threads
@@ -12,3 +14,5 @@
   - [ ] On team delete, ask user if to archive the team category
     - [ ] Add command to unarchive the team category
 - [ ] Move away from using Krypton template
+- [ ] Code cleanup 🥲
+- [ ] Test coverage
