+++ cogs/general.py
@@ -5,26 +5,17 @@ Description:
 
 Version: 6.1.0
 """
-import datetime
-import json
-import os
-import sys
-
 import platform
-import random
 
-import aiohttp
 import discord
 from discord import app_commands
 from discord.ext import commands
 from discord.ext.commands import Context
 from discord.ui import View, Select
 
-if not os.path.isfile(f"{os.path.realpath(os.path.dirname(__file__))}/../config.json"):
-    sys.exit("'config.json' not found! Please add it and try again.")
-else:
-    with open(f"{os.path.realpath(os.path.dirname(__file__))}/../config.json") as file:
-        config = json.load(file)
+from helpers import methods
+
+config = methods.load_config()
 
 
 class HelpSelect(Select):
@@ -76,7 +67,7 @@ class HelpSelect(Select):
                     inline=False
                 )
 
-        await interaction.response.send_message(embed=embed, ephemeral=True, delete_after=25)
+        await interaction.response.send_message(embed=embed, ephemeral=True)
 
 
 class General(commands.Cog, name="general"):
@@ -158,7 +149,7 @@ class General(commands.Cog, name="general"):
                 )
 
         view = View().add_item(HelpSelect(self.bot))
-        await context.send(embed=embed, view=view, delete_after=90)
+        await context.send(embed=embed, view=view, delete_after=120)
 
     # @commands.hybrid_command(
     #     name="help", description="List all commands the bot has loaded."
