+++ cogs/owner.py
@@ -7,18 +7,15 @@ Version: 6.1.0
 """
 import json
 import os
-import sys
 
 import discord
 from discord import app_commands, Forbidden, Permissions
 from discord.ext import commands
 from discord.ext.commands import Context
 
-if not os.path.isfile(f"{os.path.realpath(os.path.dirname(__file__))}/../config.json"):
-    sys.exit("'config.json' not found! Please add it and try again.")
-else:
-    with open(f"{os.path.realpath(os.path.dirname(__file__))}/../config.json") as file:
-        config = json.load(file)
+from helpers.methods import load_config
+
+config = load_config()
 
 
 class Owner(commands.Cog, name="owner"):
