+++ configs/excluded_channels.json
@@ -1,1 +1,1 @@
-[]
+[1053462751240003594, 1168267971802710026]
