+++ configs/message_info.json
@@ -1,1 +1,1 @@
-{"coaches": [{"message_id": 1159574445862686790, "channel_id": 1085592504092590080}, {"message_id": 1159591558849503324, "channel_id": 1000761876550008862}], "staff": [{"message_id": 1159591499739176960, "channel_id": 1000761876550008862}], "Alternate F4": [{"message_id": 1159994982468423770, "channel_id": 1079207596084441088}, {"message_id": 1159969264426557520, "channel_id": 1144458437678354502}, {"message_id": 1226535636815839263, "channel_id": 1082291561750069381}], "Alternate Voltage": [{"message_id": 1159995062441213952, "channel_id": 1079207596084441088}, {"message_id": 1159969887188439060, "channel_id": 1145529878435860551}], "Alternate Untitled": [{"message_id": 1159995204238057592, "channel_id": 1079207596084441088}, {"message_id": 1160327042051604532, "channel_id": 1144787996953030706}], "Alternate Inferno": [{"message_id": 1160327176349040790, "channel_id": 1079207596084441088}, {"message_id": 1160323662046560400, "channel_id": 1144477861873455124}], "Alternate Test": [{"message_id": 1161778981318688848, "channel_id": 1001057855832608808}]}
+{"coaches": [{"message_id": 1159574445862686790, "channel_id": 1085592504092590080}, {"message_id": 1159591558849503324, "channel_id": 1000761876550008862}], "staff": [{"message_id": 1159591499739176960, "channel_id": 1000761876550008862}], "Alternate F4": [{"message_id": 1159994982468423770, "channel_id": 1079207596084441088}, {"message_id": 1159969264426557520, "channel_id": 1144458437678354502}, {"message_id": 1226535370833924106, "channel_id": 1082291561750069381}], "Alternate Voltage": [{"message_id": 1159995062441213952, "channel_id": 1079207596084441088}, {"message_id": 1159969887188439060, "channel_id": 1145529878435860551}], "Alternate Untitled": [{"message_id": 1159995204238057592, "channel_id": 1079207596084441088}, {"message_id": 1160327042051604532, "channel_id": 1144787996953030706}], "Alternate Inferno": [{"message_id": 1160327176349040790, "channel_id": 1079207596084441088}, {"message_id": 1160323662046560400, "channel_id": 1144477861873455124}], "Alternate Test": [{"message_id": 1161778981318688848, "channel_id": 1001057855832608808}, {"message_id": 1230201530364199086, "channel_id": 1212005673043234816}], "Alternate Shift": [], "Alternate Dimention": [{"message_id": 1211414145425997864, "channel_id": 1204449670865100871}], "Alternate Dimension": [{"message_id": 1211640398066352228, "channel_id": 1079207596084441088}, {"message_id": 1214693162908393472, "channel_id": 1001057855832608808}, {"message_id": 1218704840859975711, "channel_id": 1001057855832608808}, {"message_id": 1218705254170755174, "channel_id": 1204449670865100871}]}
