+++ requirements.txt
@@ -1,4 +1,4 @@
-aiohttp~=3.8.5
+aiohttp~=3.8.1
 aiosqlite~=0.19.0
 python-dotenv~=1.0.0
 discord.py~=2.3.2
